Rename `name` column of `Computer` database table to `label`
All other database models already use `label` so using the same for the
`Computer` model is more consistent. The `name` attribute was already
deprecated in the ORM in `aiida-core==1.0.0` and replaced by `label`.
Here we apply the actual database migration to also perform the change
in the database models.

A new migration is also added for existing archives as the metadata
needs to update the key `name` to `label` for the `Computer` entity and
the data needs to be updated as well.