`verdi status`: add version and repository UUID to the output
The new disk objectstore-based repository provides a UUID which is also
written in the database to which it is coupled. We add it here to the
output of `verdi status` as additional information.

In addition, we add the version of the package to the output. This way,
in troubleshooting, users do not have to run `verdi --version`
separately to get this information.