`Manager`: check if repository container has been initialised
When loading the backend of the profile with the `Manager._load_backend`
method, the `get_repository_container` method of the `Profile` is
called, initialising a container for the profile in case is has not yet
been initialised. This causes issues during the migration of the
database/repository, since here the presence of the container directory
results in a critical failure of the migration command.

Here we add a check in the `_load_backend` method to see if the
container of the profile has already been initialised. If it has, the
repository uuid obtained from the container, if not it is set to `None`.
This also removes the warning that was printed during the migration,
since the repository uuid obtained from the database backend is also `None`
in before the migration.