Reinstate `aiida.manage.tests.unittest_classes`
Although this was officially deprecated and slated to be removed for
`v2.0` the deprecation warning only showed when using the `TestRunner`
class. However, many plugins only used the `PluginTestCase` run directly
through `pytest` and so didn't see the deprecation warning. Therefore,
we reinstate the module for now and will remove it in a couple of
months.