Config: warn when the configuration directory is created (#4896)
The configuration directory is automatically created when it is loaded
and it doesn't exist yet. This is useful for new installations, however,
it can lead to problems when users accidentally make a mistake when
setting an explicit path with the `AIIDA_PATH` environment variable. If
a type is present in the path, instead of the intended existing config
being used, a new empty one is created.

We add a warning to when the base configuration directory is created.
Note that this will also be emitted if the `AIIDA_PATH` variable is not
set, but even if the directory is created for the default setup this may
be useful information for the user.

The `aiida.manage.configuration.settings` module is also migrated to use
the `pathlib` module for file system path manipulations.