CLI: set `localhost` as default for database hostname in `verdi setup` (#4908)
The default was actually being defined on the option, however, it was
taken from the `pgsu.DEFAULT_DSN` dictionary, which defines the database
hostname to be `None`. Still, 9 out of 10 times the database is on the
localhost so not having this as a default is kind of annoying and
unnecessary.

Note that `pgsu` specifies `None` as the default because this is at times
the only way the psql shell can be accessed without a password.

Cherry-pick: 6c4ced3331b389cebd01a59d55eb4b07f9452672