Docs: make query docs easier to find (#4903)
With the Materials Cloud Archive adding an option to explore AiiDA
archive files on renkulab.io, finding and querying for data will be the
*first*  thing that some new AiiDA users will want to do.

Currently, the relevant documentation was hidden inside "How to work
with data", which made it difficult to find by browsing through the web
page (the page was also ranked low when searching for "query" in the
search field).

This PR makes "How to find and query for data" a top-level how-to
section. The "How to explore the provenance graph" section becomes
a subsection of "How to find and query for data", entitled "Shortcuts" 
indeed, this section deals with shortcuts in the AiiDA python API that save
users from writing QueryBuilder queries for very common generic queries.