ORM: replace `InputValidationError` with `ValueError` and `TypeError` (#4888)
The front-end ORM code threw an `InputValidationError` in a number a
places. However, this exception was intended for calculation plugins to
throw in the case of invalid input nodes. Where it was being used in the
ORM plain `ValueError` or `TypeError` exceptions should have been used.

Since the `InputValidationError` does not subclass either of the base
exception types, code that was catching this particular AiiDA-specific
exception will break after these changes. However, a scan of the most
used plugins revealed that this exception is actually not being caught
so this change should have little to no actual consequences.

In addition, the `CodeValidationError` custom exception of the class
`aiida.orm.utils.builders.code.CodeBuilder` was changed to subclass the
`ValueError` instead of the base `Exception`.