`Config`: add the `delete_profile` method (#4901)
The class already has `remove_profile` that merely removes the profile
from the index and doesn't actually delete its associated resources. The
nrem `delete_profile` method will by default also delete the repository
and database that are configured for the profile, although this behavior
can be controlled with the `include_database` and `include_repository`
arguments.

The implementation of `verdi profile delete` is update to use this new
functionality instead of the loose functions that were defined in the
`aiida.manage.configuration.setup` module. The latter, since it was not
part of the public API and was only used by this CLI command, has been
removed from the codebase.

Note that the implementation for the deletion of the database is
slightly different. Not only does `Config.delete_profile` not contain
any echoing of information, as this should be done in the CLI command
and not in the generic API, the logic for deleting the database user
has also changed. The old function `delete_database` did delete the
database user configured for the profile, as long as no other profiles
in the config used the same. This is dangerous, however, because the
same database user can be used by a program outside of AiiDA and we
don't own those users necessarily. Now, the deletion of the database
user is optional and turned off by default. When it is used though, the
database user is deleted without checking the config for other profiles
that may also have configured it.