Migrations: only show progress it not empty in file repository migration
The migration of the legacy file repository to the new disk object store
includes a progress bar counting the shards of the repository as they
are being migrated. This is useful for typical production databases for
which this can take some time, however, it was also showing for empty
databases. This output was unnecessarily noisy and on CI platforms would
give a lot of output since each iteration of the progress bar is kept
on a single line. Therefore the migration is updated to only show the
progress bar as long as the database contains at least one node.

To make this possible the code had to be refactored because if the
database is empty, now the call to `utils.migrate_legacy_repository` is
skipped entirely, since that just migrates the files of existing nodes,
however, this function was also doing other important actions, such as
initializing the new repository and checking the old repository exists,
which should definitely not be skipped. These operations are therefore
moved to the backend specific migration files themselves.

Even for empty databases, the `migrate_legacy_repository` is still
called the same as it would be for non-empty databases and it is not
skipped entirely within a simple conditional. The reason is to keep the
logic as close as possible for empty and non-empty databases since there
might be subtle differences if we skip it entirely for the one and not
the other.