Migrations: fix the rename of `name` to `label` for `DbComputer`
The migration was broken for SqlAlchemy and wasn't detected because
there were no tests. The migration of new profiles still worked because
it only failed for databases that have at least one computer.

The new migration properly uses `alter_column` to change the column
name, instead of creating a new one and dropping the old one. Not only
does this incur the exception that the new column will have null values,
which is not allowed since it is not nullable, but it would also have
lost all the data.

Finally, the `label` column has a uniqueness constraint and the name
should technically also be changed, so the constraint is dropped and
added with the new name based on `label`.