👌 IMPROVE: `Repository.hash()` speed (#4923)
The `Repository.hash` method was recursively hashing the content of all
the files it contained which is slow. However, for the disk object store
this is redundant, since the content is already hashed when it is stored
and used as the `key`. So in that case, we can actually skip the rehash
and simply return the `key`. Note that this shortcut does introduce a
dependency that the container should actually use SHA256, as that is what
should be consistent in AiiDA across all repository implementations, but
it is actually configurable. That is why a check is added to only use
the `key` if the container is configured to use SHA256.