👌 IMPROVE: Use `PurePosixPath` instead of `Path` in repository (#4930)
This makes it explicit that accessing paths in the repository is done
via posix-like paths (i.e. rather than Windows ones), and that we are
not dealing with the actual OS file-system. Note that (on Unix systems)
`Path` is a subclass of `PurePosixPath`, so you can still pass those to
the methods if you really wanted to.