`Config`: write config to disk at end of `delete_profile`
The `delete_profile` would successfully remove the profile from the
`Config` instance in memory, however, the changes would not be persisted
to disk. After reloading the config from disk, the deleted profile would
be back. A regression test is added for this behavior and the behavior
is fixed by calling `store` explictly at the end of the method, to
persist the new state to disk.

This regression was introduced in `e274ef2e065ac4aae03492b6736ae9fdea896ae4`
where the profile delete code was refactored and moved into the `Config`
class.