BaseRestartWorkChain deal with namespaced outputs
`BaseRestartWorkChain` does not return `output_namespaces` of
`_process_class` as describe in #4623. It happend because in the
`results` method of `BaseRestartWorkChain`, only output keys go through by
`node.get_outgoing` are checked and returned by the parent WorkChain.

The exposed_outputs adapt to output the whole nested namespace by
`exposed_outputs`. The `out_many` method is not used here to make a
post-check for ports to remain the original exit code check and report.