`WorkChain`: support namespacing through dots in keys of `to_context` (#4871)
It is now possible to use the `.` character in the keys used in the
keyword arguments for the `to_context` method as well as the class
`ToContext` to indicate namespaces. For example the following

    self.to_context(**{'some.name.space': future})

when the awaitable is resolved, it will retrievable from the context in

    self.ctx['some']['name']['space']

Special care needed to be taken to validate that the key represents a
nested path that can in principle have the value assigned to it. For
example, if an intermediate path in the key already has a value assigned
to it, it will not be overwritten with a dictionary to host the sub
namespace, but a `ValueError` will be raised which will cause the work
chain to except.