Transport: Avoid SSH deadlocks on large stdout/stderr (#3787)
Avoid deadlocks when retrieving stdout/stderr via SSH

We rewrite the logic by reading in chunks both stdout and stderr,
in alternating fashion (to support cases where both contain large
amounts of data).

Note that the simple shift of the line `retval = channel.recv_exit_status()`
below `stdout.read()` and `stderr.read()`, suggested in #1525, 
while partially improving the situation, was not enough
(see discussion in #1525).

In addition, the `exec_commmand_wait` functionality was significantly refactored.

Now, a `exec_commmand_wait_bytes` is defined that actually does the job,
and this needs to be implemented by plugins. 
The two plugins implemented by AiiDA now implement the new function.
Also, I improved in both the API for the stdin, so that these commands
can accept also bytes ot BytesIO objects, that makes more sense in general.

I tried to keep the API (when calling them) backward-compatible, so using e.g.
a str/StringIO as stdin still works, and similarly `exec_command_wait` works
as before, but adds an optional `encoding` parameter (the default is `utf-8` to
give the same old behavior, but give more flexibility now in using different
encodings).

However, note that the interface of plugins has changed (plugins now have to define
instead the `exec_command_wait_bytes` function instead). The change is minimal
(change the function name and avoid decoding the bytes into strings) but plugins
extending the transport functionality should take care of this (and, if they drop
the `exec_command_wait`, they should be depending on AiiDA >=2.0; otherwise they
should copy the implementation from the general Transport class if they want to remain
backward-compatible).

Also, I changed a couple of places to call directly the _bytes version, to avoid
possible issues with strange encodings on the remote computer.
Others are left in to avoid too many changes in the code, since until now noone complained
because they had a weird encoding on the supercomputer.
When such issue will arise we'll fix it, and thanks to this PR now the code is all ready
to move to treat bytes directly (or use a custom encoding, that e.g. might need to be
defined in the `verdi computer setup` or `verdi computer configure` steps, in the future).

Note on performance:
I could get ~100MB/s when not using compression, ~33MB/s when using it
(the second case is capped by the compression speed; the time is only
marginally larger of the one I get if I run, on the command line,
`time ssh -C localhost cat /path/to/remote/file | md5sum`;
the first one is not as good as running md5 (via SSH) on the command line
(which can go ~3x faster of a 4GB file) but it's still acceptable for most usecases.

(Note that here we're speaking of differences measurable only when
sending hundreds of MB sent over stdout, which should not be anyway
the default way of transferring a lot of data).

Co-authored-by: Francisco Ramirez <francisco.ramirez@epfl.ch>

Fixes #1525