`ProcessBuilder`: fix nested namespaces losing `ProcessBuilderNamespace` class
When a `ProcessBuilder` is constructed from a `PortNamespace` each
`PortNamespace` is converted into a `ProcessBuilderNamespace` including
the nested `PortNamespace`s. This guarantees that even nested port
namespaces have the same autocompletion and automatic validation when
assigning values as the top level namespace.

There was a bug however that when assigning a dictionary to a namespace
of the builder, that namespace would lose its `ProcessBuilderNamespace`
type and become a plain dictionary. Subsequent assignments to ports
within this namespace would then no longer be validated.

The source of the bug was in the `__setattr__` of `ProcessBuilderNamespace`
which would simply assign the value that is being passed to the internal
`_data` container, losing the `ProcessBuilderNamespace` class if the
value is a plain dictionary. The solution is to, in the case where an
attribute that is being set corresponds to a `PortNamespace`, to first
assign a new `ProcessBuilderNamespace` in the internal `_data` container
for that given attribute, and then iterate over the contents of the
value dictionary being assigned and call `setattr` for that on the new
`ProcessBuilderNamespace`. By calling `setattr` again, the process
described above is applied recursively, as it should.