`ProcessBuilder`: add the `_merge` method
The class already had the `_update` method, but like the normal `update`
method of a mapping, this will not recursively merge the contents of
nested dictionaries with any existing values. Anything below the top
level will simply get overwritten.

However, often one wants to merge the contents of the dictionary with
the existing namespace structure only overriding leaf nodes. To this end
the `_merge` method is added which recursively merges the content of the
new dictionary with the existing content.

Cherry-pick: 97a7fa9f4a715dec69a7d1453ae713b806425ec2