👌 IMPROVE: Allow numpy arrays to be serialized on process checkpoints (#4730)
To allow objects such as numpy arrays to be serialized to a process checkpoint,
the `AiiDALoader` now inherits from `yaml.UnsafeLoader` instead of `yaml.FullLoader`.

Note, this change represents a potential security risk,
whereby maliciously crafted code could be added to the serialized data
and then loaded upon importing an archive.
To mitigate this risk, the function `deserialize` has been renamed to `deserialize_unsafe`,
and node checkpoint attributes are removed before importing an archive.
This code is not part of the public API,
and so we assume no specific deprecations are required.

This change has also allowed for a relaxation of the `pyaml` pinning (to 5.2),
although it should be noted that this upgrade will not be realised
until a similar relaxation is implemented in plumpy.

Cherry-pick: 1bc9dbe43ff31b737ce29aca605ae15985d38ca0