Default options for codes info setting when multiple codes involved
Implementing `CalcJob` plugin with more than one code got incorrect
`codes_run_mode` option check and raise. And the `metadata.option.withmpi` was
not able to be read and set for the codes when more than one code
involved in the `CalcJob` plugin.

In this commit, the `codes_run_mode` has the default value set to `SERIAL` mode.
As for `withmpi` option, we get the value that was set in the inputs of the entire
process (by `metadata.option.withmpi`), which can then be overwritten by
the value from the `CodeInfo`. This allows plugins to force certain codes
to run without MPI, even if the user wants to run all codes with MPI whenever
possible.