Engine: only call `get_detailed_job_info` if there is a job id (#4967)
In the retrieve task of `CalcJobs` the `get_detailed_job_info` was
called always. This would lead to problems if the node did not have an
associated job id. Normally this doesn't happen because without a job id
the engine would not even have been able to confirm that the job was
ready for retrieval, however, this can happen in artificial situations
where the whole calcjob process is mocked and for example an already
completed job is passed through the system.

When there is no job id, the `get_detailed_job_info` method should not
be called because it requires the job id to get any information. Without
it, the method would except and since it is called within the
exponential backoff mechanism, the job would get stuck in the paused
state.