FIX: namespaced outputs in `BaseRestartWorkChain` (#4961)
The `BaseRestartWorkChain` did not return an `output_namespace` of
its `_process_class` as described in #4623. It happened because in its
`results` method, only the output keys are obtained from the call to
`node.get_outgoing` (checked and returned by the parent WorkChain).

This was changed for a call to `exposed_outputs`, which instead returns
the whole nested namespace. The `out_many` method is not used here 
in order to make a post-check for ports that allows to keep the original
exit code check and report.