Add the `ElectronicBandsData` data plugin
This data plugin is a subclass of the `BandsData` plugin. It adds
support to define a fermi energy, which is set as an attribute. This is
a typical use case for electronic band structures, which so far were
using the `BandsData` type, but the latter was also being used for other
band structures, such as phonons, which do not share this property.

By making it a subclass, existing processes that accept `BandsData`
instances will automatically also accept `ElectronicBandsData` instances
and queries for `BandsData` will also match the subclasses, unless the
user turns of subclassing explicitly.

The constructor would have used the `/` marker to make the `kpoints`,
`bands` and `fermi_energy` arguments positional only. The reason is that
the constructor needs to allow for arbitrary keyword arguments to be
passed to the parent constructor. This is necessary to be able to change
the `user` of the node for example. Without making the other arguments
positional only, a caller could define them as a keyword argument,
causing them to be passed to the constructor of the parent class, which
will raise an exception since it does not define those arguments.

However, we unfortunately still support Python 3.7 and `/` was not
added until Python 3.8, so for now we cannot use it yet.