Database: minimize PostgreSQL logs for Django backend (#5056)
For the Django backend, we are recording the current version of the
database schema manually in our own `DbSetting` table. We use this to
determine whether the database schema version is compatible with that of
the code that is installed.

Since the way the version and generation of the schema are stored change
over the years, we had to rely on a try-except method to determine which
way was the case. The original try was for the original way of v0.x of
`aiida-core`, and if that failed we would fall back to the modern method
of storing the info introduced in `aiida-core==1.0`. However, each time
the except was hit, meaning a programming error was encountered in the
PostgreSQL statement, a number of warning are logged to PostgreSQL. Not
only can they appear in stdout if no log file has been configured for the
PostgreSQL server (can be the case for conda installations) the log file
will also accumulate these necessary logs, which are produced every time
the database version is checked, i.e., when a profile is loaded, which
is often.

It might not be trivial to get completely rid of these errors being
logged but we can invert the try-except logic. Assuming that the
majority of users are now on v1.0 or above, they would always be hitting
the except. By first checking for the schema-style of v1.0 instead, only
users of v0.x will be getting the logs, which should be negligible.