`CalcJob`: improve default behavior for `withmpi` and `codes_run_mode` (#4990)
The `CalcJob.prepare_for_submission` contained a bug where it was
supposed to be raising an error if more than one code was specified on
the `CalcInfo` and it didn't also specify the `codes_run_mode` attribute.
However, it was incorrectly using a try-except block which would never
be hit, since the `codes_run_mode` attribute is always defined for a
`CalcInfo` instance and will simply return `None` if not explicitly
defined. Instead of fixing the check though, we change the code to simply
set `CodeRunMode.SERIAL` as the default value, which is only overridden
if the `CalcInfo` explicitly defines another value.

A similar check is done for the `CodeInfo.withmpi` attribute. This also
was required to be set by the plugin in the case that multiple codes are
to be run. Otherwise the default would be taken from the inputs of the
process instance defined in the `metadata.options.withmpi` port. Instead
here we change the logic to have that process input value to be taken as
the default, which is overridden by the value defined by the plugin in
the `CodeInfo` if and only if it is explicitly set.

There is a fundamental design issue though, because the process input
`metadata.options.withmpi` only allows a single value, which breaks down
once the user needs to be able to override specific values for a plugin
that runs multiple codes. This is a bigger design problem and is
intentionally not addressed in this commit.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
Co-authored-by: Francisco Ramirez <ramirezfranciscof@users.noreply.github.com>